Casting for character classification routines
We were being sloppy and inconsistent about casting the argument
to the character classificaton routines like isalpha(), isspace(),
isdigit(), isprint(), toupper(), tolower(), etc.

If the argument is type char, it needs to be cast to (int) to
be correct for all implementations of these library routines.

Without this we get compilation warnings on NetBSD, for
example.

+++ src/libpcp/src/config.c
@@ -37,7 +37,7 @@ dos_native_path(char *path)
 {
     char *p = path;
 
-    if (path[0] == '/' && isalpha(path[1]) && path[2] == '/') {
+    if (path[0] == '/' && isalpha((int)path[1]) && path[2] == '/') {
 	p[0] = tolower(p[1]);
 	p[1] = ':';
 	p += 2;
@@ -50,7 +50,7 @@ dos_native_path(char *path)
 static int
 dos_absolute_path(char *path)
 {
-    return (isalpha(path[0]) && path[1] == ':' && path[2] == '\\');
+    return (isalpha((int)path[0]) && path[1] == ':' && path[2] == '\\');
 }
 
 static char *
@@ -59,7 +59,7 @@ msys_native_path(char *path)
     char *p = path;
 
     /* Only single drive letters allowed (Wikipedia says so) */
-    if (isalpha(path[0]) && path[1] == ':') {
+    if (isalpha((int)path[0]) && path[1] == ':') {
 	p[1] = tolower(p[0]);
 	p[0] = '/';
 	p += 2;
